Recursively process the VEX expressions into shadow-state computations.
For now, to vastly simplify the implementation, the primops are *all*
being handled by slow unsafe calls into the Rust helper.

I still have not quite figured out the right way to get the borrow-mut
client request to be registered as the initial introducer of the shadow-tag.
That is priority numero uno for next hacking session.