added shadow-map tagging for temps.
I am observing a shadow value being attached to temp 9 and
subsequently that shadow value being loaded on a subsequent read of
temp 9. So, that is progress. (I don't see the shadow value flow
further from that point, but that's probably because I haven't
implemented the flow for arithemtic operations yet.)