Fixed big bug in the borrow_mut handler: you put a tag with the counter, not with the address. Facepalm.
Added some code to try to check the stack associated with an address
and pop the entries as dictated by stacked borrowed model. (This is
what uncovered the aforementioned bug.)

Added an `alert` macro that has same interface as `msg`. Intent is to
allow me to disable the msg output easily as I approach the goal post
here.

Added method to convert a stack entry to a num, just for ease of
printing, though that will probably need to be revised or just
disappear once I add more interesting entry variants.