fixed the known bug with the return-value in rs_shadow_load once I realized that the new client-req protocol was relying on it functioning.
Extended STACKED_BORROW_EVENT to hold the stash address as well as the borrowed
address itself, and then used that to validate the flow of the new protocol as
it runs. (So now we no longer rely on the hack for our values, but we are double
checking our values via the hack, for now.)

Added extra conditionalization of the instrumentation near msg invocations so
that we won't see random IRop output in the midst of our run unless you have
turned PRINT_MSG to true.