Review feedback: We can use Rust's formatting machinery rather than hard code every call to vgPlain_umsg.
The actual review feedback has suggested using `fmt::Debug` here, but I opted
not to do that, because I intepret these umsg calls as being part of the
user-visible output from the tool, and for that, `fmt::Display` is the
appropriate trait, not `fmt::Debug`. (This is a shame, because there are no
great ways (at least not in libstd) to recursively print list/set/map
collections using the elements' `fmt::Display` implementations. But I am not
going to try to tackle that problem today; I just wrote the standard code to
iterate over the list.

The thing that *was* fun about this code: I added `VgPlainUmsgWriter`, which is
a `fmt::Writer` implementation that emits output via `vgPlain_usmg`. This does
simplify things nicely, and I suspect we can use it to clean other other parts
of the code base.